Use authority instead of tenantName and tenantId
This commit fixes an assumption that the authority will always be a combination of the `tenantId` and `tenantName`.
Unfortunately this is not true for Azure B2C tenants that are using custom user flows. Each user flow has a unique
well-known configuration URL that includes a parameter `p` specifying the name of the user flow. Each user flow
returns a unique ID token. The exclusion of the `p` parameter results in a `404` from Azure B2C.